Fixes issues 127, 130, and 132.
127: The issue was in the block if nbands > 1: -- in these cases,
these are sources with GaiaPositions AND nbands == 1. We are not fitting positions but this block was never entered because nbands was 1 and thus we have 2 extra columns of zeros in our Xic. This leads to the shape mismatch when computing the priors arrays.

Solution: if nbands >= 1.

130: len(tr.images) is actually 0 in these cases so we should call the
     CPU version instead where this case is already handled.

132: There are modelMasks that are None.  This is handled cleanly in the
     CPU version so call it when any modelMask is None because easier to
handle there than in batch mode.