Update to fix bug 136, which is the same as bug 121 but for FFT not MoGs.
            #IF2 = IF[nfft > 0]
            #flipped_mask = IF2.sum(2, keepdims=1) > cp.arange(self.maxNfft-1,-1,-1)
            flipped_mask = IF.sum(2, keepdims=1) > cp.arange(self.maxNfft-1,-1,-1)

Repeating what I wrote for the mogs version, I can't remember why I had subscripted only where nmogs > 0 (here nfft > 0) but things seem to run happily so far without that and the array shapes were wrong on this particular source when I did do that subscript. Perhaps it was something that had just not been tested and was a mistaken assumption on my part?